move mqtt fixture to conftest.py, so that it will be available for multiple test files. SEE: https://docs.pytest.org/en/stable/reference/fixtures.html\#conftest-py-sharing-fixtures-across-multiple-files
